Discovery Only mode
Stops scoreReadClass from running in quant only mode.

+++ R/bambu.R
@@ -85,9 +85,15 @@
 bambu <- function(reads = NULL, rcFile = NULL, rcOutDir = NULL,
     annotations = NULL, genome = NULL, stranded = FALSE, ncore = 1,
     yieldSize = NULL, opt.discovery = NULL, opt.em = NULL,
-    discovery = TRUE, verbose = FALSE) {
+    discoveryOnly = FALSE, quantOnly = FALSE, verbose = FALSE) {
+    if (discoveryOnly & quantOnly) {
+        warning(paste0("At least 1 of discoveryOnly and quantOnly must be FALSE. 
+ Rerun with either 1 or both parameters as FALSE"))
+        break
+    }
     annotations <- checkInputs(annotations, reads, readClass.file = rcFile, 
             readClass.outputDir = rcOutDir, genomeSequence = genome)
+    genomeSequence <- checkInputSequence(genome)
     isoreParameters <- setIsoreParameters(isoreParameters = opt.discovery)
     emParameters <- setEmParameters(emParameters = opt.em)
     bpParameters <- setBiocParallelParameters(reads, readClass.file = rcFile,
@@ -103,31 +109,41 @@ bambu <- function(reads = NULL, rcFile = NULL, rcOutDir = NULL,
             rm.readClassSe <- TRUE # remove temporary read class files 
         }
         readClassList <- bambu.processReads(reads, annotations, 
-            genomeSequence = genome, 
+            genomeSequence = genomeSequence, 
             readClass.outputDir = rcOutDir, yieldSize, 
-            bpParameters, stranded, verbose,
-            min.readCount = isoreParameters[["min.readCount"]])
+            bpParameters, stranded, verbose)
     } else {
         readClassList <- rcFile
     }
-    if (discovery) {
+    if (!quantOnly) {
+        readClassList <- bplapply(seq_along(readClassList), function(i) {
+            scoreReadClasses(readClassList[[i]],genomeSequence, 
+                             annotations, 
+                             min.readCount = isoreParameters[["min.readCount"]], 
+                             verbose = verbose)},
+            BPPARAM = bpParameters)
         annotations <- bambu.extendAnnotations(readClassList, annotations,
             isoreParameters, stranded, bpParameters, verbose = verbose)
         if (!verbose) message("Finished extending annotations.")
+        if (discoveryOnly){
+            return(annotations)
+        }
     }
-    if (!verbose) message("Start isoform quantification")
-    countsSe <- bplapply(readClassList,
-        bambu.quantify,annotations = annotations,
-        min.exonDistance = isoreParameters[["min.exonDistance"]],
-        min.primarySecondaryDist =
-            isoreParameters[['min.primarySecondaryDist']], 
-        min.primarySecondaryDistStartEnd =
-            isoreParameters[['min.primarySecondaryDistStartEnd2']],
-        emParameters = emParameters, ncore = ncore,
-        verbose = verbose, BPPARAM = bpParameters)
-    countsSe <- do.call(SummarizedExperiment::cbind, countsSe)
-    rowRanges(countsSe) <- annotations
-    if (!verbose) message("Finished isoform quantification.")
-    if (rm.readClassSe) file.remove(unlist(readClassList))#Clean temp directory
-    return(countsSe)
+    if (!discoveryOnly) {
+        if (!verbose) message("Start isoform quantification")
+        countsSe <- bplapply(readClassList,
+            bambu.quantify,annotations = annotations,
+            min.exonDistance = isoreParameters[["min.exonDistance"]],
+            min.primarySecondaryDist =
+                isoreParameters[['min.primarySecondaryDist']], 
+            min.primarySecondaryDistStartEnd =
+                isoreParameters[['min.primarySecondaryDistStartEnd2']],
+            emParameters = emParameters, ncore = ncore,
+            verbose = verbose, BPPARAM = bpParameters)
+        countsSe <- do.call(SummarizedExperiment::cbind, countsSe)
+        rowRanges(countsSe) <- annotations
+        if (!verbose) message("Finished isoform quantification.")
+        if (rm.readClassSe) file.remove(unlist(readClassList))#Clean temp directory
+        return(countsSe)
+    }
 }
